Bring ActiveStorage code to overriden controllers
Bring the latest code from ActiveStorage gem to our overriden
controllers.
Some of the changes are precisely comments about how to add additional
security to activestorage access, that was the cause of our controllers
override (so we can add user access control).

It may be worth to take a look at the newly introduced documentation, as
seems there is a simpler Rails configuration way to disable the default
direct file routes, instead of overriding the controller as we have done
to disable them.