Restrict Coverband to production environment
We only want it running in Review Apps/Staging/Production, not on
development/testing.

Theoretically this can be disabled with an ENV variable, but coverband
in Rails keeps loading before the env vas get loaded in local
environment, causing issues with local Redis and not being disabled on
command.

The solution: Only load the gem in production environment.