Mark unused views for future deletion (#2787)
* Mark app error summary page for deletion

As the partial is not loaded by any view. The error summary used from
our service is loaded from GovUKDesignSystem gem.

* Mark "app/_registered_product" view as unused

It is not called by any view in the service.

* Mark "app/_unfinished_product" view as unused

As it is not rendered by any view in the service.

* Mark old confirmation pages for deletion

They do not match any controller, and devise confirmation pages are
either inside the gem or rewritten in our service under a different
route (that matches our service controller).

* Mark PoisonC. trigger rules page for deletion

As this page is duplicated but unused, the used one is on
/notifications/, not /notifications_search/ path.

* Mark incompl. notifications accordion for deletion

This accordion view is not used since we moved to the draft notifications
page.

Also deleting a related unused feature spec helper.

* Mark "poisonous ingredients" page for removal

Not loaded from anywhere. Replaced by
"contains_ingredients_npis_needs_to_know" page.

* Mark old Sign Up page for deletion

This page is neither used for Submit or Search users registration.