Mark ruby code as unused
After scraping over all the files in Coverband production report and
testing and discarding possible false positives, we are marking the
methods and files identified as unreachable/unused by the service.

We have not included code logic branches that are usually not hit but
still reachable/needed for the service, but current
dangling/unused/uncalled methods and files.