Make TriggerQuestion relationships non-optional
Both the relationship between TriggerQuestion - Component and
TriggerQuestion - TriggerQuestionElement where marked as optional with
some TODO to enforce the relationship removing the optionality once the
CPNP parser has been fixed/improved.

We do not use the parser anymore (it got removed when we stopped allowing
CPNP Zip file imports) and these relationships should be enforced now.

I kind of infer that there was some issue with missing imported
information from the CPNP files back in the days, but this is not the
case anymore.