Cleanup orphaned Trigger Question data (#2801)
* Make TriggerQuestion relationships non-optional

Both the relationship between TriggerQuestion - Component and
TriggerQuestion - TriggerQuestionElement where marked as optional with
some TODO to enforce the relationship removing the optionality once the
CPNP parser has been fixed/improved.

We do not use the parser anymore (it got removed when we stopped allowing
CPNP Zip file imports) and these relationships should be enforced now.

I kind of infer that there was some issue with missing imported
information from the CPNP files back in the days, but this is not the
case anymore.

* Add task to cleanup dangling trigger questions

We have TriggerQuestions in DB not associated with any Component, so
they're effectively not associated with any Notification or any other
model. They are unusable/unreachable.

Same for TriggerQuestionElements that are not associated with any
TriggerQuestion.

We are deleting them with a rake task.

* Fix rake task further tests not running the task

The tests were passing because they were asserting that something wasn't
happening, but underneath the rake task wasn't running after the first
test was ran. This code reenables the task between tests so it will run
per each test.

* Introduce alerts for dangling data loaded from DB

If any of the affected TriggerQuestion or TriggerQuestionElement gets
loaded from DB, it will send an alert to Sentry.

* Reword dangling as orphaned

Better word for the concept we are describing.