Highlight whole ingredient names, not just the hit
ES highlighting only highlights the hit term, but not the whole
ingredient name if its longer than the searching term that caused the
hit.

We want to highlight the whole name of the ingredient. So we are using
the highlight marking tags as a way to identify which ingredients from
the list have been causing the search hit.

Once isolated, we locally remove the partial highlighting tags and wrap
the whole ingredient name with a <mark> tag.