Only validate "used_for_multiple_shades" for exact ingredients
The form does not display the selector for this value when filling a
range concentration ingredient.

The validation was blocking the form from being submitted with range
concentration ingredients.

On the same way, the model validation was blocking cloning
notifications with range ingredients, since the
"used_for_multiple_shades" isn't filled for those and fails validation
for the new records.

Solution: Make the validation optional, as we do not want the value
being set for range ingredients.