Do not change notified timestamp when unarchiving
Since the archiving was setting the notification to a new state from
notification_complete, and unarchive set it up back to
notification_complete, the default recording of the state timestamp by
the AASM gem was overriding the initial "notification_complete_at"
value.

We want to keep the original date when the notification was submitted,
and we can achieve that by disabling AASM automated
notification_complete_at recording, and setting it as a callback
executed exclusively when the notification moves from draft to
completion.