Bugfix: Fix exception when uploading ingredients file with extra columns (#2877)
* Fix exception when ingredient CSV has extra columns

While parsing the ingredient rows from the CSV file into individual
ingredients, the parser method was raising an exception if the row has
more columns than the expected ones. As we use a strict list of
arguments for parsing.

Resolved by allowing any extra arguments when calling the parser. If any
extra arguments are provided, the parser will return without attempting
to build an Ingredient.

Based on that return, we will be able to return the error on the
particular line.

If this happens on the header we will just ignore it, as long as the
ingredient rows are ok, we do not care about what extra info may come in
the header row.

* Refactor: Inline assigning component

Just to save the extra call/line.

---------

Co-authored-by: Ruben Arakelyan <ruben@arakelyan.uk>