Add notification archiving (#2873)
Adds functionality to enable responsible persons to archive notifications for one of a number of set reasons. Archived notifications are displayed separately but are otherwise identical to non-archived ones. Archived notifications can also be unarchived at any time. A history is kept of each archival/unarchival along with the selected reason and this is displayed to any team member of the RP.

Co-authored-by: Marc Sardón <marc@bichito.software>