Recovery code improvements (#2976)
* Redirect user if already logged in

Checks if a user is already logged in before accepting a recovery code, and redirects them if so. This prevents inadvertent use of multiple recovery codes if users return to the entry page using the browser back button.

* Strip whitespace from recovery codes

Removes all whitespace from entered recovery codes. This is because recovery codes are displayed in two parts with a space in between for readability. Entering a code in this format would result in a validation error.

* Remove unnecessary page furniture from recovery code print view

Prevents various buttons and text from being printed when a user chooses to print their recovery codes.

* Remove recovery codes feature flag

This feature has been successfully rolled out to all users.

* Test recovery codes with spaces